requirements.txt
Django>=4.2.11,<5
psycopg[binary]==3.1.18

asgiref==3.8.0

feedparser==6.0.8

Markdown==3.6

smartypants==2.0.1
olefile==0.47

versiontools==1.9.1
statsd==4.0.1
pep8==1.7.1
pyflakes==3.2.0
mccabe==0.7.0
pyparsing==3.1.0
pycodestyle==2.11.0
configparser==7.0.0
importlib-metadata<7.2 # for flake8
typing_extensions==4.11.0
zipp==3.18.0  # for flake8
flake8==7.0.0
tornado==6.4
cssselect==1.2.0
lxml==5.2.0
fuzzywuzzy==0.18.0
sure==2.0.0

coverage==7.4.0
pyasn1==0.6.0
cffi==1.16.0 # cryptography
cryptography==42.0.4  # pyOpenSSL


chardet==5.2.0
idna==3.7
urllib3==2.2.1  # sentry-sdk

requests==2.31.0
logilab-common==2.0.0
lazy-object-proxy==1.10.0
wrapt==1.16.0

pylint==3.1.0

astroid==3.1.0

six==1.16.0
text-unidecode==1.3
Faker==24.11.0
factory_boy==3.3.0
ldap3==2.9.1
sqlparse==0.5.0
contextlib2==21.6.0
rcssmin==1.1.0
rjsmin==1.2.0

vine==5.1.0
amqp==5.2.0
kombu==5.3.4
prompt-toolkit==3.0.32
celery==5.3.6
backports.zoneinfo;python_version<"3.9"  # celery
click==8.1.3
click-didyoumean==0.3.0
click-repl==0.3.0
click-plugins==1.1.1
django-celery-results==2.5.0
wcwidth==0.2.5

billiard==4.2.0
pytz==2024.1

django-indexer==0.3.0
django-templatetag-sugar==1.0
django-paging==0.2.5
django-appconf==1.0.5
django-statsd-mozilla==0.4.0
raven==6.10.0
sentry-sdk==1.45.0
django-bootstrap-form==3.4
django-debug-toolbar==4.3.0
django-waffle==4.1.0
django-smoketest==1.2.0
django-extensions==3.2.0

django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-impersonate==1.9.1
django-registration-redux==2.13
gunicorn==22.0.0
future==1.0.0
s3transfer==0.10.0
jmespath==1.0.0

botocore>=1.15.0,<1.35.0
boto3>=1.12.0,<1.35.0

certifi==2024.2.2  # for pysolr, sentry-sdk

pysolr==3.9.0
django-haystack==3.2.1
requirements/src/django_audit_log-0.8.2-py2.py3-none-any.whl
djangorestframework==3.14.0
djangorestframework-jsonp==1.0.2
Unidecode==1.3.2
django-reversion==5.0.1
python-dateutil==2.9.0
requirements/src/edtf-0.9.3ctl-py2.py3-none-any.whl

django-storages==1.14.2
ctlsettings==0.3.2
funcsigs==1.0.2
pbr==6.0.0
django-s3sign<=0.2.1
django-smtp-ssl==1.0

pbr==6.0.0
pyyaml==6.0.1
stevedore>=1.20.0 # Apache-2.0
bandit==1.7.0

entrypoints==0.4
distro==1.9.0

xlwt==1.3.0
django-adminactions==2.3.0

uritemplate==4.1.1
jsonschema==4.21.0
inflection==0.5.1
drf-spectacular==0.27.0

python-cas==1.6.0
django-cas-ng==5.0.0
